bhoutgate/requirements.txt
PySide6==6.9.0
Flask==2.3.3
paho-mqtt==1.6.1
Werkzeug==2.3.7 